Address PR feedback and harden websocket/pubsub handling
